chore(typing): clean 26 mypy --strict errors in 12 modules
Resolve all outstanding mypy --strict errors across core/kuramoto/,
core/physics/, and tacl/ (69 source files now clean).

Files (12):
- core/kuramoto/{adaptive,delayed,early_stopping,second_order,sparse}.py
- core/physics/{diffusion_predictor,engine,gravitational_coupling,
  maxwell,relativity,thermodynamic_risk,wave_propagation}.py

Fixes applied:
- np.ndarray without parameters -> NDArray[np.float64]
- Returning Any from typed functions -> explicit float()/cast()
- list[Any] from .tolist() on indices -> list comprehension with int()
- Parameterize generic dict/list/tuple
- No # type: ignore[...] used; no public API signatures changed

Quality gates:
- mypy --strict on core/kuramoto/ + core/physics/ + tacl/: clean
- ruff check + ruff format --check: clean on all 12 files
- black --check: clean on all 12 files
- pytest tests/unit/core/ tests/unit/physics/ tests/tacl/: pass
  (1 pre-existing skip unrelated to typing changes)

+++ core/physics/thermodynamic_risk.py
@@ -95,7 +95,7 @@ class ThermodynamicRiskGate:
         if total < 1e-12:
             return 0.0
         w = w / total
-        return (1.0 - float(np.sum(w ** self._q))) / (self._q - 1.0)
+        return (1.0 - float(np.sum(w**self._q))) / (self._q - 1.0)
 
     def ricci_temperature(self, kappa_min: float) -> float:
         """T_eff = T_base · exp(-κ_min).
@@ -103,7 +103,7 @@ class ThermodynamicRiskGate:
         κ_min < 0 (negative curvature) → T_eff > T_base → looser risk.
         κ_min > 0 (positive curvature) → T_eff < T_base → tighter risk.
         """
-        return self._T_base * np.exp(-kappa_min)
+        return float(self._T_base * np.exp(-kappa_min))
 
     @staticmethod
     def free_energy(U: float, T: float, S: float) -> float:
